Completed requested changes.
- Changed popover title from "Info" to "Lab Types".
- Moved popover code to a separate component and called it from LabEditor.tsx.
