+++ src/pages/LabEditor/LabEditor.tsx
@@ -27,9 +27,7 @@ import {ButtonLink} from '../../components/util/ButtonLink';
 import {FileInput} from '../../components/util/FileInput';
 import {VmTemplateModal} from '../../components/VmTemplateModal/VmTemplateModal';
 import {BridgeListEditor} from '../../components/BridgeListEditor/BridgeListEditor';
-import {FontAwesomeIcon} from '@fortawesome/react-fontawesome';
-import { faInfoCircle } from '@fortawesome/free-solid-svg-icons';
-import {OverlayTrigger, Popover} from 'react-bootstrap';
+import {LabTypes} from '../../components/LabTypesPopover/LabTypesPopover';
 
 const labDifficultyOptions: DropdownOption<LabDifficulty>[] = [
   {value: 1, label: 'Easy'},
@@ -106,26 +104,7 @@ export default function LabEditor({match: {params: {moduleId, labId}}}: Props) {
   }, [labId]);
 
   const getFieldName = (name: keyof LabForm) => name;
-
   const [selectedVm, setSelectedVm] = useState<number|undefined>();
-  const popover = (
-    <Popover id='popover-basic' style={{maxWidth: '750px'}}>
-      <Popover.Title as='h3'>Info:</Popover.Title>
-      <Popover.Content>
-        <div>- Temporary - only persist for only an hour and torn down afterwards. Their is no long term persistence with temporary labs.</div>
-        <div>
-          - Semester - these labs persist until a certain date-time is reached. 
-          Once the date-time is reach the lab is torn down is unable to be started again.
-        </div>
-        <div>- Permanent - these labs are never torn down.</div>
-      </Popover.Content>
-    </Popover>
-  );
-  const Info = () => (
-    <OverlayTrigger trigger='hover' placement='right' overlay={popover}>
-      <a><FontAwesomeIcon icon={faInfoCircle}/></a>
-    </OverlayTrigger>
-  );
   const renderForm = () => (
     <Formik
       initialValues={initialValues}
@@ -167,7 +146,7 @@ export default function LabEditor({match: {params: {moduleId, labId}}}: Props) {
                    {values.hasReadme && <a href={getUserLabReadmeUrl(values.id)} target='_blank'>View Current</a>}
                  </Form.Group>
                  <Form.Group>
-                   <Form.Label>Lab Type <Info /></Form.Label>
+                   <Form.Label>Lab Type <LabTypes /></Form.Label>
                    <DropdownInput name={getFieldName('type')} dropdownData={labTypeOptions} disabled={!editing}/>
                  </Form.Group>
                  <Form.Group>
